move order monitor prioritization tests
